add unit test for subdividing a merged face
