--- Debug/Editor/TempMenuItems.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: d4261a5057f18944ba81f792a283c52b
-timeCreated: 1458917851
-licenseType: Pro
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
