+++ Runtime/Core/AssemblyInfo.cs
@@ -1,6 +1,7 @@
 [assembly: System.Runtime.CompilerServices.InternalsVisibleTo("Unity.ProBuilder.MeshOperations")]
 [assembly: System.Runtime.CompilerServices.InternalsVisibleTo("Unity.ProBuilder.Editor")]
 [assembly: System.Runtime.CompilerServices.InternalsVisibleTo("Unity.ProBuilder.Tests")]
+[assembly: System.Runtime.CompilerServices.InternalsVisibleTo("Unity.ProBuilder.Tests.Framework")]
 [assembly: System.Runtime.CompilerServices.InternalsVisibleTo("Unity.ProBuilder.Tests.Slow")]
 [assembly: System.Runtime.CompilerServices.InternalsVisibleTo("Unity.ProBuilder.Editor.Tests")]
 [assembly: System.Runtime.CompilerServices.InternalsVisibleTo("Unity.ProBuilder.Debug.Editor")]
